+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-beta.21] - 2026-03-16
+
+### Added
+
+- `DeviceCapabilityModule` with `isEsimSupported()` support, exported response/result types, unit tests, and manual documentation
+- `UserAttributesModule` with `getSelectedTravelDestination()` support, exported response/result types, unit tests, and manual documentation
+
 ## [2.0.0-beta.20] - 2026-03-16
 
 ### Added